Reorganize
`EffectID` can be defined anywhere and at any point of time, but they should only be accessed when a `Reducer` is processing an action. A runtime warning is raised otherwise.
