+++ Sources/ComposableArchitecture/Effects/Cancellation.swift
@@ -104,71 +104,3 @@ extension Effect {
 
 var cancellationCancellables: [AnyHashable: Set<AnyCancellable>] = [:]
 let cancellablesLock = NSRecursiveLock()
-
-@propertyWrapper
-public struct EffectID: Hashable {
-  let effectIdentifier: EffectIdentifier
-
-  public var wrappedValue: EffectIdentifier {
-    effectIdentifier
-  }
-
-  static var currentContextID: AnyHashable? {
-    Thread.current.threadDictionary.value(forKey: currentContextKey) as? AnyHashable
-  }
-
-  public init<ID>(
-    wrappedValue: ID,
-    file: String = #file,
-    line: UInt = #line,
-    column: UInt = #column
-  ) where ID: Hashable {
-    effectIdentifier = .init(
-      contextID: Self.currentContextID,
-      userData: wrappedValue,
-      file: file,
-      line: line,
-      column: column
-    )
-  }
-
-  public init(
-    file: String = #file,
-    line: UInt = #line,
-    column: UInt = #column
-  ) {
-    effectIdentifier = .init(
-      contextID: Self.currentContextID,
-      file: file,
-      line: line,
-      column: column
-    )
-  }
-}
-
-public struct EffectIdentifier: Hashable {
-  init(
-    contextID: AnyHashable? = nil,
-    userData: AnyHashable? = nil,
-    file: String = #file,
-    line: UInt = #line,
-    column: UInt = #column
-  ) {
-    self.contextID = contextID
-    self.userData = userData
-    self.file = file
-    self.line = line
-    self.column = column
-    #if DEBUG
-      if contextID == nil {
-        // TODO: Log some warning about the id not being store-specific?
-      }
-    #endif
-  }
-
-  let contextID: AnyHashable?
-  let userData: AnyHashable?
-  let file: String
-  let line: UInt
-  let column: UInt
-}
